requirements.txt
numpy==1.21.0
opencv-python==4.5.2.54
Pillow==8.3.1
